desenvolvimento_fullstack_django_vuejs/paris_cafe/requirements.txt
Django
djangorestframework
markdown
django-filter
django-cors-headers
pytest
pytest-django
pytest-html
model_bakery